+++ src/app/users/login/login-form.tsx
@@ -1,16 +1,14 @@
 'use client';
 
+import styles from '@/styles/app/users/login/page.module.scss';
 import { InputString } from '@/src/components/common/form/input-string';
 import { InputPassword } from '@/src/components/common/form/input-password';
 import { FormEvent, useState } from 'react';
 import { Button } from '@/src/components/common/button';
 import { useUserContext } from '@/src/contexts/user.context';
 import { doUserLogin } from '@/src/api/api';
-
-import styles from '@/styles/app/users/login/page.module.scss';
 import { toastError, toastSuccess } from '@/src/utils/toast.utils';
 import { Loader } from '@/src/components/common/loader';
-import { BackLink } from '@/src/components/common/back-link';
 import Link from 'next/link';
 
 export function LoginForm() {
@@ -42,7 +40,7 @@ export function LoginForm() {
     return (
         <form onSubmit={onSubmit}>
             {isLogging && <Loader isAbsolute={true} />}
-            <BackLink link="/" label={'Powrót do strony głównej'} isAttached={true} />
+            {/*<BackLinkBar link="/" label={'Powrót do strony głównej'} isAttached={true} />*/}
             <div>
                 <div>
                     <h3>Zaloguj się</h3>